+++ packages/create-app/src/enums/template.ts
@@ -8,7 +8,10 @@ export enum Template {
   ReactTypeScript = 'react-ts',
   ReactJavaScript = 'react',
   ReactTypeScriptWebpack = 'react-ts-webpack',
+  ReactViteJavaScript = 'react-vite',
+  ReactViteTypeScript = 'react-vite-ts',
   MainOnlyTypeScript = 'main-ts',
   MainOnlyJavaScript = 'main',
   SvelteJavaScript = 'svelte',
+  SvelteTypeScript = 'svelte-ts',
 }