+++ packages/create-app/templates/react-vite-ts/src/renderer/vite.config.js
@@ -1,0 +1,10 @@
+const { defineConfig } = require('vite')
+const reactRefresh = require('@vitejs/plugin-react-refresh')
+
+// https://vitejs.dev/config/
+module.exports = defineConfig({
+  plugins: [reactRefresh()],
+  build: {
+    target: 'chrome89',
+  },
+})